fix warnings an more

+++ led_str.c
@@ -41,7 +41,7 @@ int led_regex_match(pcre2_code* regex, const char* line, int len) {
     return rc > 0;
 }
 
-int led_regex_match_offset(pcre2_code* regex, const char* line, int len, int* pzone_start, int* pzone_stop) {
+int led_regex_match_offset(pcre2_code* regex, const char* line, int len, size_t* pzone_start, size_t* pzone_stop) {
     pcre2_match_data* match_data = pcre2_match_data_create_from_pattern(regex, NULL);
     int rc = pcre2_match(regex, (PCRE2_SPTR)line, len, 0, 0, match_data, NULL);
     led_debug("match_offset %d ", rc);
